Cleaned up page testing by creating new mockingForPageTests class
